+++ app/build.gradle.kts
@@ -6,11 +6,11 @@ plugins {
 }
 
 android {
-    namespace = "com.example.first_week_mission"
+    namespace = "com.kky.pokedex"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.first_week_mission"
+        applicationId = "com.kky.pokedex"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
