--- app/src/main/java/com/example/first_week_mission/GlobalApplication.kt
@@ -1,9 +1,0 @@
-package com.example.first_week_mission
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class GlobalApplication: Application() {
-
-}
