--- app/src/main/java/com/example/first_week_mission/di/PokemonRepositoryModule.kt
@@ -1,17 +1,0 @@
-package com.example.first_week_mission.di
-
-import com.example.first_week_mission.repository.PokemonRepository
-import com.example.first_week_mission.repository.PokemonRepositoryImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal abstract class PokemonRepositoryModule {
-    @Singleton
-    @Binds
-    abstract fun bindPokemonRepository(pokemonRepositoryImpl: PokemonRepositoryImpl): PokemonRepository
-}
