--- app/src/main/java/com/example/first_week_mission/di/PokemonServiceModule.kt
@@ -1,42 +1,0 @@
-package com.example.first_week_mission.di
-
-import com.example.first_week_mission.service.PokemonApiService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal object PokemonServiceModule {
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(): OkHttpClient {
-        val builder = OkHttpClient.Builder()
-
-        val logging = HttpLoggingInterceptor()
-        logging.setLevel(HttpLoggingInterceptor.Level.BODY)
-        builder.addInterceptor(logging)
-
-        return builder.build()
-    }
-
-    @Provides
-    @Singleton
-    fun providePokemonService(okHttpClient: OkHttpClient): PokemonApiService {
-        val baseUrl = "https://pokeapi.co/api/v2/"
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl(baseUrl)
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-        return  retrofit.create(PokemonApiService::class.java)
-    }
-}
