--- app/src/main/java/com/example/first_week_mission/domain/model/PokemonSummary.kt
@@ -1,10 +1,0 @@
-package com.example.first_week_mission.domain.model
-
-data class PokemonSummary(
-    override val id: Int,
-    override val name: String,
-    override val description: String,
-    override val type: List<String>,
-    override val like: Boolean,
-    override val imageUrl: String
-) : Pokemon
