--- app/src/main/java/com/example/first_week_mission/domain/model/Stat.kt
@@ -1,6 +1,0 @@
-package com.example.first_week_mission.domain.model
-
-data class Stat(
-    val baseStat: Int,
-    val name: String
-)
