--- app/src/main/java/com/example/first_week_mission/model/FlavorText.kt
@@ -1,8 +1,0 @@
-package com.example.first_week_mission.model
-
-import com.google.gson.annotations.SerializedName
-
-data class FlavorText(
-    @SerializedName("flavor_text") val flavorText: String,
-    val language: Info
-)
