--- app/src/main/java/com/example/first_week_mission/model/Info.kt
@@ -1,6 +1,0 @@
-package com.example.first_week_mission.model
-
-data class Info(
-    val name: String,
-    val url: String
-)
