--- app/src/main/java/com/example/first_week_mission/model/PokemonDetailResponse.kt
@@ -1,20 +1,0 @@
-package com.example.first_week_mission.model
-
-import com.google.gson.annotations.SerializedName
-
-internal data class PokemonResponse(
-    val id: Long,
-    val stats: List<StatInfo>,
-    val abilities: List<AbilityInfo>,
-    val weight: Int,
-    val height: Int
-)
-
-internal data class StatInfo(
-    @SerializedName("base_stat") val baseStat:Int,
-    @SerializedName("stat") val stat: Info
-)
-
-internal data class AbilityInfo(
-    val ability: Info
-)
