--- app/src/main/java/com/example/first_week_mission/model/StatResponse.kt
@@ -1,11 +1,0 @@
-package com.example.first_week_mission.model
-
-data class StatResponse(
-    val name: String,
-    val names: List<StatName>
-)
-
-data class StatName(
-    val language: Info,
-    val name: String
-)
