+++ app/src/main/java/com/kky/pokedex/GlobalApplication.kt
@@ -1,0 +1,9 @@
+package com.kky.pokedex
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class GlobalApplication: Application() {
+
+}
