+++ app/src/main/java/com/kky/pokedex/di/AppDatabaseModule.kt
@@ -1,0 +1,25 @@
+package com.kky.pokedex.di
+
+import android.content.Context
+import androidx.room.Room
+import com.kky.pokedex.local.AppDatabase
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppDatabaseModule {
+    @Singleton
+    @Provides
+    fun getAppDatabase(@ApplicationContext context: Context): AppDatabase {
+        return Room.databaseBuilder(
+            context,
+            AppDatabase::class.java,
+            "app-database"
+        ).build()
+    }
+}
