+++ app/src/main/java/com/kky/pokedex/di/PokemonRepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.kky.pokedex.di
+
+import com.kky.pokedex.repository.PokemonRepository
+import com.kky.pokedex.repository.PokemonRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class PokemonRepositoryModule {
+    @Singleton
+    @Binds
+    abstract fun bindPokemonRepository(pokemonRepositoryImpl: PokemonRepositoryImpl): PokemonRepository
+}
