+++ app/src/main/java/com/kky/pokedex/domain/model/Stat.kt
@@ -1,0 +1,6 @@
+package com.kky.pokedex.domain.model
+
+data class Stat(
+    val baseStat: Int,
+    val name: String
+)
