+++ app/src/main/java/com/kky/pokedex/model/AbilityResponse.kt
@@ -1,0 +1,7 @@
+package com.kky.pokedex.model
+
+import com.google.gson.annotations.SerializedName
+
+data class AbilityResponse(
+    @SerializedName("flavor_text_entries") val flavorTexts: List<FlavorText>
+)
