+++ app/src/main/java/com/kky/pokedex/model/FlavorText.kt
@@ -1,0 +1,8 @@
+package com.kky.pokedex.model
+
+import com.google.gson.annotations.SerializedName
+
+data class FlavorText(
+    @SerializedName("flavor_text") val flavorText: String,
+    val language: Info
+)
