+++ app/src/main/java/com/kky/pokedex/model/Info.kt
@@ -1,0 +1,6 @@
+package com.kky.pokedex.model
+
+data class Info(
+    val name: String,
+    val url: String
+)
