+++ app/src/main/java/com/kky/pokedex/model/StatResponse.kt
@@ -1,0 +1,11 @@
+package com.kky.pokedex.model
+
+data class StatResponse(
+    val name: String,
+    val names: List<StatName>
+)
+
+data class StatName(
+    val language: Info,
+    val name: String
+)
